coordinator: move peerrecovery from enterprise
